pyproject.toml
[project]
name = "receipt-convert"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.166.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "openai>=1.70.0",
    "pypdf[image]>=5.4.0",
]

[tool.uv]
package = true

[project.scripts]
main = "main:main"
